data: update fake enrollment data
The dataset needed `term_desc` and `days_to_class_start` columns before
it could be used in the app.

`days_to_class_start` is identical to `days_to_term_start` except that
it has been filtered to be < 21 (see recent changes in the
daily_enrollment.sql file in scheduled-pins).

The stored dataset has been nested on `days_to_class_start` to limit the
amount of duplication for other columns for a given student/term. This
reduced a 8M row dataset to a 100k row dataset.

Overall the .rds file has reduced in size from 8.3MB to 1.5MB.